Move all the parsing logic out of internal model
