+++ pipeline/transformer/transformer_common/src/main/scala/weco/pipeline/transformer/parse/PeriodParser.scala
@@ -1,4 +1,4 @@
-package weco.catalogue.internal_model.parse
+package weco.pipeline.transformer.parse
 
 import fastparse._
 import NoWhitespace._
