+++ pipeline/transformer/transformer_common/src/main/scala/weco/pipeline/transformer/transformers/ParsedPeriod.scala
@@ -1,16 +1,17 @@
 package weco.pipeline.transformer.transformers
 
-import weco.catalogue.internal_model.text.TextNormalisation._
 import weco.catalogue.internal_model.identifiers.IdState
-import weco.catalogue.internal_model.work.{InstantRange, Period}
-import weco.catalogue.internal_model.parse.parsers.DateParser
+import weco.catalogue.internal_model.text.TextNormalisation._
+import weco.catalogue.internal_model.work.Period
+import weco.pipeline.transformer.parse.PeriodParser
 
 object ParsedPeriod {
   def apply(label: String): Period[IdState.Unidentifiable.type] = {
     val normalisedLabel = label.trimTrailingPeriod
     Period(
-      IdState.Unidentifiable,
-      normalisedLabel,
-      InstantRange.parse(normalisedLabel))
+      id = IdState.Unidentifiable,
+      label = normalisedLabel,
+      range = PeriodParser(normalisedLabel)
+    )
   }
 }
