+++ pipeline/transformer/transformer_common/src/test/scala/weco/pipeline/transformer/parse/PeriodParserTest.scala
@@ -1,4 +1,4 @@
-package weco.catalogue.internal_model.parse
+package weco.pipeline.transformer.parse
 
 import java.time.LocalDate
